feat!: Enhanced `AsyncStreamCDC`
- Support both `tokio` and `futures` via feature flags
- Add documentation and usage examples for `AsyncStreamCDC`
- Add tests that specifically target `AsyncStreamCDC`
- Run both `tokio` and `futures`-based tests in CI
