+++ Cargo.toml
@@ -13,6 +13,12 @@ exclude = [
     "test/*",
 ]
 
+[features]
+default = ["tokio"]
+tokio = ["dep:tokio", "tokio-stream", "async-stream"]
+futures = ["dep:futures"]
+
+
 [dev-dependencies]
 aes = "0.8.2"
 byteorder = "1.4.3"
@@ -20,6 +26,12 @@ clap = { version = "4.2.1", features = ["cargo"] }
 ctr = "0.9.2"
 md-5 = "0.10.5"
 memmap2 = "0.5.8"
+tokio = { version = "1.29.1", features = ["io-util", "rt", "macros"] }
+futures-test = { version = "0.3.28" }
 
 [dependencies]
-futures = "0.3.28"
+futures = { version = "0.3.28", optional = true }
+tokio = { version = "1.29.1", features = ["io-util"], optional = true }
+tokio-stream = { version = "0.1.14", optional = true }
+async-stream = { version = "0.3.5", optional = true }
+
