+++ src/v2020/mod.rs
@@ -34,9 +34,8 @@
 use std::fmt;
 use std::io::Read;
 
-use futures::io::AsyncRead;
-use futures::io::AsyncReadExt;
-use futures::stream::Stream;
+mod async_stream_cdc;
+pub use async_stream_cdc::*;
 
 /// Smallest acceptable value for the minimum chunk size.
 pub const MINIMUM_MIN: u32 = 64;
@@ -59,7 +58,7 @@ pub const MAXIMUM_MAX: u32 = 16_777_216;
 // the deduplication ratio is slightly improved when the mask bits are spread
 // relatively evenly, hence these seemingly "magic" values.
 //
-const MASKS: [u64; 26] = [
+pub(self) const MASKS: [u64; 26] = [
     0,                  // padding
     0,                  // padding
     0,                  // padding
@@ -239,7 +238,7 @@ const GEAR_LS: [u64; 256] = [
 
 // Find the next chunk cut point in the source.
 #[allow(clippy::too_many_arguments)]
-fn cut(
+pub(self) fn cut(
     source: &[u8],
     min_size: usize,
     avg_size: usize,
@@ -317,7 +316,7 @@ pub enum Normalization {
 }
 
 impl Normalization {
-    fn bits(&self) -> u32 {
+    pub(self) fn bits(&self) -> u32 {
         match self {
             Normalization::Level0 => 0,
             Normalization::Level1 => 1,
@@ -711,162 +710,10 @@ impl<R: Read> Iterator for StreamCDC<R> {
     }
 }
 
-pub struct AsyncStreamCDC<R> {
-    /// Buffer of data from source for finding cut points.
-    buffer: Vec<u8>,
-    /// Maximum capacity of the buffer (always `max_size`).
-    capacity: usize,
-    /// Number of relevant bytes in the `buffer`.
-    length: usize,
-    /// Source from which data is read into `buffer`.
-    source: R,
-    /// Number of bytes read from the source so far.
-    processed: u64,
-    /// True when the source produces no more data.
-    eof: bool,
-    min_size: usize,
-    avg_size: usize,
-    max_size: usize,
-    mask_s: u64,
-    mask_l: u64,
-    mask_s_ls: u64,
-    mask_l_ls: u64,
-}
-
-impl<R: AsyncRead + std::marker::Unpin> AsyncStreamCDC<R> {
-    ///
-    /// Construct a `StreamCDC` that will process bytes from the given source.
-    ///
-    /// Uses chunk size normalization level 1 by default.
-    ///
-    pub fn new(source: R, min_size: u32, avg_size: u32, max_size: u32) -> Self {
-        Self::with_level(source, min_size, avg_size, max_size, Normalization::Level1)
-    }
-
-    ///
-    /// Create a new `StreamCDC` with the given normalization level.
-    ///
-    pub fn with_level(
-        source: R,
-        min_size: u32,
-        avg_size: u32,
-        max_size: u32,
-        level: Normalization,
-    ) -> Self {
-        assert!(min_size >= MINIMUM_MIN);
-        assert!(min_size <= MINIMUM_MAX);
-        assert!(avg_size >= AVERAGE_MIN);
-        assert!(avg_size <= AVERAGE_MAX);
-        assert!(max_size >= MAXIMUM_MIN);
-        assert!(max_size <= MAXIMUM_MAX);
-        let bits = logarithm2(avg_size);
-        let normalization = level.bits();
-        let mask_s = MASKS[(bits + normalization) as usize];
-        let mask_l = MASKS[(bits - normalization) as usize];
-        Self {
-            buffer: vec![0_u8; max_size as usize],
-            capacity: max_size as usize,
-            length: 0,
-            source,
-            eof: false,
-            processed: 0,
-            min_size: min_size as usize,
-            avg_size: avg_size as usize,
-            max_size: max_size as usize,
-            mask_s,
-            mask_l,
-            mask_s_ls: mask_s << 1,
-            mask_l_ls: mask_l << 1,
-        }
-    }
-
-    /// Fill the buffer with data from the source, returning the number of bytes
-    /// read (zero if end of source has been reached).
-    async fn fill_buffer(&mut self) -> Result<usize, Error> {
-        // this code originally copied from asuran crate
-        if self.eof {
-            Ok(0)
-        } else {
-            let mut all_bytes_read = 0;
-            while !self.eof && self.length < self.capacity {
-                let bytes_read = self.source.read(&mut self.buffer[self.length..]).await?;
-                if bytes_read == 0 {
-                    self.eof = true;
-                } else {
-                    self.length += bytes_read;
-                    all_bytes_read += bytes_read;
-                }
-            }
-            Ok(all_bytes_read)
-        }
-    }
-
-    /// Drains a specified number of bytes from the buffer, then resizes the
-    /// buffer back to `capacity` size in preparation for further reads.
-    fn drain_bytes(&mut self, count: usize) -> Result<Vec<u8>, Error> {
-        // this code originally copied from asuran crate
-        if count > self.length {
-            Err(Error::Other(format!(
-                "drain_bytes() called with count larger than length: {} > {}",
-                count, self.length
-            )))
-        } else {
-            let data = self.buffer.drain(..count).collect::<Vec<u8>>();
-            self.length -= count;
-            self.buffer.resize(self.capacity, 0_u8);
-            Ok(data)
-        }
-    }
-
-    /// Find the next chunk in the source. If the end of the source has been
-    /// reached, returns `Error::Empty` as the error.
-    async fn read_chunk(&mut self) -> Result<ChunkData, Error> {
-        self.fill_buffer().await?;
-        if self.length == 0 {
-            Err(Error::Empty)
-        } else {
-            let (hash, count) = cut(
-                &self.buffer[..self.length],
-                self.min_size,
-                self.avg_size,
-                self.max_size,
-                self.mask_s,
-                self.mask_l,
-                self.mask_s_ls,
-                self.mask_l_ls,
-            );
-            if count == 0 {
-                Err(Error::Empty)
-            } else {
-                let offset = self.processed;
-                self.processed += count as u64;
-                let data = self.drain_bytes(count)?;
-                Ok(ChunkData {
-                    hash,
-                    offset,
-                    length: count,
-                    data,
-                })
-            }
-        }
-    }
-
-    pub fn as_stream(&mut self) -> impl Stream<Item = Result<ChunkData, Error>> + '_ {
-        futures::stream::unfold(self, |this| async {
-            let chunk = this.read_chunk().await;
-            if let Err(Error::Empty) = chunk {
-                None
-            } else {
-                Some((chunk, this))
-            }
-        })
-    }
-}
-
 ///
 /// Base-2 logarithm function for unsigned 32-bit integers.
 ///
-fn logarithm2(value: u32) -> u32 {
+pub(self) fn logarithm2(value: u32) -> u32 {
     f64::from(value).log2().round() as u32
 }
